fix: update reviews e votes services

+++ insis-review-command/src/main/java/com/isep/acme/services/implementations/ReviewServiceImpl.java
@@ -1,14 +1,17 @@
 package com.isep.acme.services.implementations;
 
 import java.time.LocalDate;
-
+import java.util.HashSet;
+import java.util.Set;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.isep.acme.controllers.ResourceNotFoundException;
+import com.isep.acme.dtos.VoteReviewDTO;
 import com.isep.acme.mappers.ReviewMapper;
 import com.isep.acme.model.Product;
 import com.isep.acme.model.Review;
+import com.isep.acme.model.Vote;
 import com.isep.acme.repositories.ProductRepository;
 import com.isep.acme.repositories.ReviewRepository;
 import com.isep.acme.services.ProductService;
@@ -17,9 +20,12 @@ import com.isep.acme.services.RestService;
 import com.isep.acme.services.ReviewService;
 
 import lombok.AllArgsConstructor;
+import lombok.extern.java.Log;
+import lombok.extern.slf4j.Slf4j;
 
 @Service
 @AllArgsConstructor
+@Slf4j
 public class ReviewServiceImpl implements ReviewService {
 
     @Autowired
@@ -48,14 +54,13 @@ public class ReviewServiceImpl implements ReviewService {
         review.setFunFact(funfact);
         review.setProduct(product);
         review.setApprovalStatus("pending");
-
         reviewRepository.save(review);
 
         return review;
     }
 
     @Override
-    public Review create(Review review) {
+    public Review createWithReview(Review review) {
 
         reviewRepository.findById(review.getIdReview())
                 .ifPresent(existingReview -> {
@@ -70,8 +75,8 @@ public class ReviewServiceImpl implements ReviewService {
             throws ResourceNotFoundException, IllegalArgumentException {
 
         Review reviewToUpdate = reviewRepository.findById(reviewID)
-                .orElseThrow(() -> new ResourceNotFoundException("Review not found"));
-
+            .orElseThrow(() -> new ResourceNotFoundException("Review not found"));
+        System.out.println("Review to update: " + reviewToUpdate.getVotes());
         if (reviewToUpdate.setApprovalStatus(status))
             return reviewRepository.save(reviewToUpdate);
 
@@ -83,8 +88,28 @@ public class ReviewServiceImpl implements ReviewService {
 
         Review reviewToDelete = reviewRepository.findById(reviewId)
                 .orElseThrow(() -> new ResourceNotFoundException("Review not found"));
-
+        //not possible delet review with votes
+        if (reviewToDelete.getVotes() != null){
+            log.info("Review with votes can't be deleted");
+            return false;
+        }
+   
         reviewRepository.delete(reviewToDelete);
         return true;
+    }
+
+    @Override
+    public Review addVoteToReview(Long reviewID, Vote vote){
+        try {
+            Review review = reviewRepository.findById(reviewID).orElseThrow(() -> new ResourceNotFoundException("Review not found"));        
+            System.out.println("Review: " + review.getVotes());
+            review.addVote(vote);
+            System.out.println("Review: " + review.getVotes());
+            return reviewRepository.save(review);
+        } catch (Exception e) {
+            e.printStackTrace();
+            throw new IllegalArgumentException("Error adding vote to review");
+        }
+
     }
 }
